+++ src/Hal/Tests/Serializer/CollectionNormalizerTest.php
@@ -131,24 +131,24 @@ class CollectionNormalizerTest extends TestCase
             $paginator->method('count')->willReturn(12);
         }
 
-        $resourceClassResolverMock = $this->createMock(ResourceClassResolverInterface::class);
-        $resourceClassResolverMock->method('getResourceClass')->with($paginator, 'Foo')->willReturn('Foo');
+        $resourceClassResolverStub = $this->createStub(ResourceClassResolverInterface::class);
+        $resourceClassResolverStub->method('getResourceClass')->willReturnMap([[$paginator, 'Foo', 'Foo']]);
 
-        $resourceMetadataFactoryMock = $this->createMock(ResourceMetadataCollectionFactoryInterface::class);
-        $resourceMetadataFactoryMock->method('create')->with('Foo')->willReturn(new ResourceMetadataCollection('Foo', [
+        $resourceMetadataFactoryStub = $this->createStub(ResourceMetadataCollectionFactoryInterface::class);
+        $resourceMetadataFactoryStub->method('create')->willReturnMap([['Foo', new ResourceMetadataCollection('Foo', [
             (new ApiResource())->withShortName('Foo')->withOperations(new Operations([
                 'bar' => (new GetCollection())->withShortName('Foo'),
             ])),
-        ]));
+        ])]]);
 
         $itemNormalizer = $this->createMock(NormalizerInterface::class);
-        $itemNormalizer->method('normalize')->with('foo', CollectionNormalizer::FORMAT, [
+        $itemNormalizer->expects($this->once())->method('normalize')->with('foo', CollectionNormalizer::FORMAT, [
             'resource_class' => 'Foo',
             'api_sub_level' => true,
             'root_operation_name' => 'bar',
         ])->willReturn(['_links' => ['self' => '/me'], 'name' => 'Kévin']);
 
-        $normalizer = new CollectionNormalizer($resourceClassResolverMock, 'page', $resourceMetadataFactoryMock);
+        $normalizer = new CollectionNormalizer($resourceClassResolverStub, 'page', $resourceMetadataFactoryStub);
         $normalizer->setNormalizer($itemNormalizer);
 
         return $normalizer->normalize($paginator, CollectionNormalizer::FORMAT, [