+++ src/Symfony/Bundle/Resources/config/symfony/events.php
@@ -64,6 +64,7 @@ return static function (ContainerConfigurator $container) {
         ->args([
             service('api_platform.state_provider.parameter.inner'),
             tagged_locator('api_platform.parameter_provider', 'key'),
+            service('api_platform.pagination_options'),
         ]);
 
     $services->set('api_platform.listener.request.read', ReadListener::class)